+++ Vapour.Server.Host/Controller/Endpoints/ControllerHostStatusEndpoint.cs
@@ -1,9 +1,8 @@
 ﻿using FastEndpoints;
 
 using Vapour.Server.Controller;
-using Vapour.Server.Host.Controller;
 
-namespace Vapour.Server.Host.Endpoints;
+namespace Vapour.Server.Host.Controller.Endpoints;
 
 public class ControllerHostStatusEndpoint : EndpointWithoutRequest<ControllerHostStatusResponse>
 {
